Rename submodules under `cupyx.scipy.special`
