+++ cupyx/scipy/special/__init__.py
@@ -1,27 +1,27 @@
-from cupyx.scipy.special.bessel import i0  # NOQA
-from cupyx.scipy.special.bessel import i1  # NOQA
-from cupyx.scipy.special.bessel import j0  # NOQA
-from cupyx.scipy.special.bessel import j1  # NOQA
-from cupyx.scipy.special.bessel import y0  # NOQA
-from cupyx.scipy.special.bessel import y1  # NOQA
+from cupyx.scipy.special._bessel import i0  # NOQA
+from cupyx.scipy.special._bessel import i1  # NOQA
+from cupyx.scipy.special._bessel import j0  # NOQA
+from cupyx.scipy.special._bessel import j1  # NOQA
+from cupyx.scipy.special._bessel import y0  # NOQA
+from cupyx.scipy.special._bessel import y1  # NOQA
 
-from cupyx.scipy.special.polygamma import polygamma  # NOQA
+from cupyx.scipy.special._polygamma import polygamma  # NOQA
 
-from cupyx.scipy.special.digamma import digamma  # NOQA
-from cupyx.scipy.special.gamma import gamma  # NOQA
-from cupyx.scipy.special.gammaln import gammaln  # NOQA
-from cupyx.scipy.special.zeta import zeta  # NOQA
+from cupyx.scipy.special._digamma import digamma  # NOQA
+from cupyx.scipy.special._gamma import gamma  # NOQA
+from cupyx.scipy.special._gammaln import gammaln  # NOQA
+from cupyx.scipy.special._zeta import zeta  # NOQA
 
-from cupyx.scipy.special.statistics import ndtr  # NOQA
+from cupyx.scipy.special._statistics import ndtr  # NOQA
 
-from cupyx.scipy.special.erf import erf  # NOQA
-from cupyx.scipy.special.erf import erfc  # NOQA
-from cupyx.scipy.special.erf import erfcx  # NOQA
-from cupyx.scipy.special.erf import erfinv  # NOQA
-from cupyx.scipy.special.erf import erfcinv  # NOQA
+from cupyx.scipy.special._erf import erf  # NOQA
+from cupyx.scipy.special._erf import erfc  # NOQA
+from cupyx.scipy.special._erf import erfcx  # NOQA
+from cupyx.scipy.special._erf import erfinv  # NOQA
+from cupyx.scipy.special._erf import erfcinv  # NOQA
 
-from cupyx.scipy.special.convex_analysis import entr  # NOQA
-from cupyx.scipy.special.convex_analysis import huber  # NOQA
-from cupyx.scipy.special.convex_analysis import kl_div  # NOQA
-from cupyx.scipy.special.convex_analysis import pseudo_huber  # NOQA
-from cupyx.scipy.special.convex_analysis import rel_entr  # NOQA
+from cupyx.scipy.special._convex_analysis import entr  # NOQA
+from cupyx.scipy.special._convex_analysis import huber  # NOQA
+from cupyx.scipy.special._convex_analysis import kl_div  # NOQA
+from cupyx.scipy.special._convex_analysis import pseudo_huber  # NOQA
+from cupyx.scipy.special._convex_analysis import rel_entr  # NOQA
