+++ cupyx/scipy/special/_convex_analysis.py
