+++ cupyx/scipy/special/_digamma.py
